[AN] chore: CarouselCard 컴포저블 분리 #777
